Refactor bpf metric collect package
