+++ pkg/profiler/cpu/bpfmetrics/collector_test.go
@@ -10,7 +10,8 @@
 // WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 // See the License for the specific language governing permissions and
 // limitations under the License.
-package cpu
+
+package bpfmetrics
 
 import (
 	"testing"
